Struct aws_sdk_lexruntime::client::fluent_builders::PutSession

pub struct PutSession<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> PutSession<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<PutSessionOutput, SdkError<PutSessionError>> where
R::Policy: SmithyRetryPolicy<PutSessionInputOperationOutputAlias, PutSessionOutput, PutSessionError, PutSessionInputOperationRetryAlias>,

pub fn bot_name(self, inp: impl Into<String>) -> Self

The name of the bot that contains the session data.

pub fn set_bot_name(self, input: Option<String>) -> Self

pub fn bot_alias(self, inp: impl Into<String>) -> Self

The alias in use for the bot that contains the session data.

pub fn set_bot_alias(self, input: Option<String>) -> Self

pub fn user_id(self, inp: impl Into<String>) -> Self

The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot.

pub fn set_user_id(self, input: Option<String>) -> Self

pub fn session_attributes(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

Adds a key-value pair to sessionAttributes.

To override the contents of this collection use set_session_attributes.

Map of key/value pairs representing the session-specific context information. It contains application information passed between Amazon Lex and a client application.

pub fn set_session_attributes(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn dialog_action(self, inp: DialogAction) -> Self

Sets the next action that the bot should take to fulfill the conversation.

pub fn set_dialog_action(self, input: Option<DialogAction>) -> Self

pub fn recent_intent_summary_view(self, inp: impl Into<IntentSummary>) -> Self

Appends an item to recentIntentSummaryView.

To override the contents of this collection use set_recent_intent_summary_view.

A summary of the recent intents for the bot. You can use the intent summary view to set a checkpoint label on an intent and modify attributes of intents. You can also use it to remove or add intent summary objects to the list.

An intent that you modify or add to the list must make sense for the bot. For example, the intent name must be valid for the bot. You must provide valid values for:

intentName

slot names

slotToElict

If you send the recentIntentSummaryView parameter in a PutSession request, the contents of the new summary view replaces the old summary view. For example, if a GetSession request returns three intents in the summary view and you call PutSession with one intent in the summary view, the next call to GetSession will only return one intent.

pub fn set_recent_intent_summary_view(
self,
input: Option<Vec<IntentSummary>>
) -> Self

pub fn accept(self, inp: impl Into<String>) -> Self

The message that Amazon Lex returns in the response can be either text or speech based depending on the value of this field.

If the value is text/plain; charset=utf-8, Amazon Lex returns text in the response.

If the value begins with audio/, Amazon Lex returns speech in the response. Amazon Lex uses Amazon Polly to generate the speech in the configuration that you specify. For example, if you specify audio/mpeg as the value, Amazon Lex returns speech in the MPEG format.

If the value is audio/pcm, the speech is returned as audio/pcm in 16-bit, little endian format.

The following are the accepted values:

audio/mpeg

audio/ogg

audio/pcm

audio/* (defaults to mpeg)

text/plain; charset=utf-8

pub fn set_accept(self, input: Option<String>) -> Self

pub fn active_contexts(self, inp: impl Into<ActiveContext>) -> Self

Appends an item to activeContexts.

To override the contents of this collection use set_active_contexts.

A list of contexts active for the request. A context can be activated when a previous intent is fulfilled, or by including the context in the request,

If you don't specify a list of contexts, Amazon Lex will use the current list of contexts for the session. If you specify an empty list, all contexts for the session are cleared.

pub fn set_active_contexts(self, input: Option<Vec<ActiveContext>>) -> Self