Module aws_sdk_lexruntime::error

Modules

bad_gateway_exception

See BadGatewayException

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

dependency_failed_exception

See DependencyFailedException

internal_failure_exception

See InternalFailureException

limit_exceeded_exception

See LimitExceededException

loop_detected_exception

See LoopDetectedException

not_acceptable_exception

See NotAcceptableException

not_found_exception

See NotFoundException

request_timeout_exception

See RequestTimeoutException

unsupported_media_type_exception

See UnsupportedMediaTypeException

Structs

BadGatewayException

Either the Amazon Lex bot is still building, or one of the dependent

BadRequestException

Request validation failed, there is no usable message in the context,

ConflictException

Two clients are using the same AWS account, Amazon Lex bot, and user

DeleteSessionError

DependencyFailedException

One of the dependencies, such as AWS Lambda or Amazon Polly, threw an

GetSessionError

InternalFailureException

Internal service error. Retry the call.

LimitExceededException

Exceeded a limit.

LoopDetectedException

This exception is not used.

NotAcceptableException

The accept header in the request does not have a valid value.

NotFoundException

The resource (such as the Amazon Lex bot or an alias) that is referred

PostContentError

PostTextError

PutSessionError

RequestTimeoutException

The input speech is too long.

UnsupportedMediaTypeException

The Content-Type header (PostContent API) has an invalid

Enums

DeleteSessionErrorKind

GetSessionErrorKind

PostContentErrorKind

PostTextErrorKind

PutSessionErrorKind