Module aws_sdk_lexruntime::model

Modules

active_context

See ActiveContext

active_context_time_to_live

See ActiveContextTimeToLive

button

See Button

dialog_action

See DialogAction

generic_attachment

See GenericAttachment

intent_confidence

See IntentConfidence

intent_summary

See IntentSummary

predicted_intent

See PredictedIntent

response_card

See ResponseCard

sentiment_response

See SentimentResponse

Structs

ActiveContext

A context is a variable that contains information about the current

ActiveContextTimeToLive

The length of time or number of turns that a context remains

Button

Represents an option to be shown on the client platform (Facebook,

DialogAction

Describes the next action that the bot should take in its interaction

GenericAttachment

Represents an option rendered to the user when a prompt is shown. It

IntentConfidence

Provides a score that indicates the confidence that Amazon Lex has that an

IntentSummary

Provides information about the state of an intent. You can use this

PredictedIntent

An intent that Amazon Lex suggests satisfies the user's intent. Includes

ResponseCard

If you configure a response card when creating your bots, Amazon Lex

SentimentResponse

The sentiment expressed in an utterance.

Enums

ConfirmationStatus

ContentType

DialogActionType

DialogState

FulfillmentState

MessageFormatType